chore(componens) renamed compontent
