+++ src/heads-up/components/collapsible-list.vue
